// packages/EditorTools/enum/point-type.ts
/** 枚举位置 */
export enum POINT_TYPE {
  LEFT_TOP = 'leftTop',
  RIGHT_TOP = 'rightTop',
  LEFT_BOTTOM = 'leftBottom',
  RIGHT_BOTTOM = 'rightBottom',
  TOP_CENTER = 'topCenter',
  RIGHT_CENTER = 'rightCenter',
  BOTTOM_CENTER = 'bottomCenter',
  LEFT_CENTER = 'leftCenter',
  ROTATE = 'rotate',
}